fix(nexus-web): refactor event fetching to use direct API calls instead of proxy route (#1216)
